# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "thaeeithantun.dev@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "L5 Swagger OpenApi Server"
        }
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "get users ",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "User Id",
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "description": "User Name",
                                                        "type": "string"
                                                    },
                                                    "email": {
                                                        "description": "User Email",
                                                        "type": "string"
                                                    },
                                                    "role": {
                                                        "description": "User Role",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi"
                                        },
                                        "message": {
                                            "description": "Message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create user with form data",
                "operationId": "createuserWithForm",
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Enter name",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Enter email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Enter password",
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "description": "Confirm Password",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Enter Role",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "User Id",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "User Name",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "User Email",
                                            "type": "string"
                                        },
                                        "role": {
                                            "description": "User Role",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/users/{userId}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user information",
                "description": "Returns user data",
                "operationId": "getuserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "User Id",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "User Name",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "User Email",
                                            "type": "string"
                                        },
                                        "role": {
                                            "description": "User Role",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Updates user with form data",
                "operationId": "updateuserWithForm",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID of user that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Updated email of the user",
                                        "type": "string"
                                    },
                                    "role": {
                                        "description": "Updated role of the user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete user",
                "description": "Delete user",
                "operationId": "deleteuserById",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "user Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ],
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}